+++ data-transfer/src/test/java/it/eng/datatransfer/exceptions/DataTransferExceptionAdviceTest.java
@@ -32,6 +32,14 @@ class DataTransferExceptionAdviceTest {
 	}
 	
 	@Test
+	public void handleTransferProcessNotFoundException() {
+		TransferProcessNotFoundException ex = new TransferProcessNotFoundException(TEST_ERROR_MESSAGE);
+		ResponseEntity<Object> response = advice.handleTransferProcessNotFoundException(ex, request);
+		assertNotNull(response);
+		assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());
+	}
+	
+	@Test
 	public void handleTransferProcessExistsException() {
 		TransferProcessExistsException ex = new TransferProcessExistsException(TEST_ERROR_MESSAGE, ModelUtil.CONSUMER_PID);
 		ResponseEntity<Object> response = advice.handleTransferProcessExistsException(ex, request);